+++ src/Reed/Classes/AppDelegate.swift
@@ -75,7 +75,7 @@ final class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationC
         AppDefaults.registerDefaults()
 
         registerBackgroundTasks()
-        CacheCleaner.purgeIfNecessary()
+        purgeCachesIfNecessary()
         initializeDownloaders()
         initializeHomeScreenQuickActions()
 
@@ -237,6 +237,42 @@ final class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationC
 // MARK: App Initialization
 
 extension AppDelegate {
+    private func purgeCachesIfNecessary() {
+        guard let flushDate = AppDefaults.shared.lastImageCacheFlushDate else {
+            AppDefaults.shared.lastImageCacheFlushDate = Date()
+            return
+        }
+
+        // If the image disk cache hasn't been flushed for 3 days and the network is available, delete it
+        if flushDate.addingTimeInterval(3600 * 24 * 3) < Date() {
+            if NetworkMonitor.shared.isConnected {
+                let tempDir = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!
+                let faviconsFolderURL = tempDir.appendingPathComponent("Favicons")
+                let imagesFolderURL = tempDir.appendingPathComponent("Images")
+                let feedURLToIconURL = tempDir.appendingPathComponent("FeedURLToIconURLCache.plist")
+                let homePageToIconURL = tempDir.appendingPathComponent("HomePageToIconURLCache.plist")
+                let homePagesWithNoIconURL = tempDir.appendingPathComponent("HomePagesWithNoIconURLCache.plist")
+
+                for tempItem in [
+                    faviconsFolderURL,
+                    imagesFolderURL,
+                    feedURLToIconURL,
+                    homePageToIconURL,
+                    homePagesWithNoIconURL,
+                ] {
+                    do {
+                        DZLog("Removing cache file: \(tempItem.absoluteString)")
+                        try FileManager.default.removeItem(at: tempItem)
+                    } catch {
+                        DZLog("Could not delete cache file: \(error.localizedDescription)")
+                    }
+                }
+
+                AppDefaults.shared.lastImageCacheFlushDate = Date()
+            }
+        }
+    }
+
     private func initializeDownloaders() {
         let tempDir = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!
         let imagesFolderURL = tempDir.appendingPathComponent("Images")