+++ src/Reed/Classes/ExtensionContainersFile+MainApp.swift
@@ -81,3 +81,24 @@ extension ExtensionContainersFile {
         }
     }
 }
+
+extension ExtensionDataStore {
+    @MainActor
+    init(dataStore: DataStore) {
+        self.name = dataStore.nameForDisplay
+        self.dataStoreID = dataStore.dataStoreID
+        self.disallowFeedInRootFolder = false
+        self.containerID = dataStore.containerID
+        self.folders = dataStore.sortedFolders?.map { ExtensionFolder(folder: $0) } ?? [ExtensionFolder]()
+    }
+}
+
+extension ExtensionFolder {
+    @MainActor
+    init(folder: Folder) {
+        self.dataStoreName = folder.dataStore?.nameForDisplay ?? ""
+        self.dataStoreID = folder.dataStore?.dataStoreID ?? ""
+        self.name = folder.nameForDisplay
+        self.containerID = folder.containerID
+    }
+}